# docs/onboarding_swagger.yaml
openapi: 3.0.0
info:
  title: Onboarding APIs
  version: 1.0.0
components:
  securitySchemes:
paths:
...
  /api/v1/platform/account:
    post:
      tags:
        - Account Management
      summary: Create Account
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accountname:
                  type: string
                  example: Test Account 1
                mobile:
                  type: string
                  example: "9848312193"
                accountinfo:
                  type: object
                  example: {}
                isenabled:
                  type: boolean
                  example: true
              required:
                - accountname
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/v1/platform/model/:
    post:
      tags:
        - Vehicle Models
      summary: Create Model
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                modelcode:
                  type: string
                  example: treo_default
                  description: modelcode = trim(lowercase(modelname)) + "_" + trim(lowercase(modelvariant))
                modelname:
                  type: string
                  example: Treo
                modelvariant:
                  type: string
                  example: Default
                modelfamilycode:
                  type: string
                  example: treo
                modeldisplayname:
                  type: string
                  example: TREO Default
              required:
                - modelcode
                - modelname
                - modelvariant
                - modelfamilycode
                - modeldisplayname
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...